+++ packages/validation/src/validators/definitions/CommandDefinitions.ts
@@ -1317,7 +1317,7 @@ export const CommandDefinitions: MessageGroup = {
                     mayBeNull: false,
                 },
                 {
-                    name: "newTarget",
+                    name: "newReference",
                     type: "LionWebId",
                     isList: false,
                     isOptional: true,
@@ -1378,7 +1378,7 @@ export const CommandDefinitions: MessageGroup = {
                     mayBeNull: false,
                 },
                 {
-                    name: "deletedTarget",
+                    name: "deletedReference",
                     type: "LionWebId",
                     isList: false,
                     isOptional: true,
@@ -1439,7 +1439,7 @@ export const CommandDefinitions: MessageGroup = {
                     mayBeNull: false,
                 },
                 {
-                    name: "oldTarget",
+                    name: "oldReference",
                     type: "LionWebId",
                     isList: false,
                     isOptional: true,
@@ -1453,7 +1453,7 @@ export const CommandDefinitions: MessageGroup = {
                     mayBeNull: true,
                 },
                 {
-                    name: "newTarget",
+                    name: "newReference",
                     type: "LionWebId",
                     isList: false,
                     isOptional: true,
@@ -1490,7 +1490,7 @@ export const CommandDefinitions: MessageGroup = {
             ],
         },
         {
-            name: "ChunkedCommand",
+            name: "ContinuedCommand",
             properties: [
                 {
                     name: "chunk",