+++ packages/class-core/src/deltas/appliers.ts
@@ -347,6 +347,7 @@ const deltaApplier = (idMapping?: IdMapping, updatablePartitions?: () => INodeBa
             }
             if (delta instanceof CompositeDelta) {
                 delta.parts.forEach(applyDelta);
+                return;
             }
             if (delta instanceof NoOpDelta) {
                 return;