+++ packages/validation/src/validators/definitions/EventDefinitions.ts
@@ -622,13 +622,6 @@ export const EventDefinitions: MessageGroup = {
                     mayBeNull: false,
                 },
                 {
-                    name: "split",
-                    type: "Boolean",
-                    isList: false,
-                    isOptional: true,
-                    mayBeNull: false,
-                },
-                {
                     name: "messageKind",
                     type: "String",
                     isList: false,
@@ -997,7 +990,7 @@ export const EventDefinitions: MessageGroup = {
                     mayBeNull: false,
                 },
                 {
-                    name: "newIndex",
+                    name: "indexOffset",
                     type: "Number",
                     isList: false,
                     isOptional: false,
@@ -1487,7 +1480,7 @@ export const EventDefinitions: MessageGroup = {
                     mayBeNull: false,
                 },
                 {
-                    name: "newIndex",
+                    name: "indexOffset",
                     type: "Number",
                     isList: false,
                     isOptional: false,