+++ packages/validation/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## 0.10.0
 
+* Change key values pairs for additional info's to a dictionary and update
+the validator to be able to validate the dictionary. 
 * Update to latest spec. of delta protocol.
 Especially additional properties for all Move/Replace child and annotation commands.