+++ packages/validation/CHANGELOG.md
@@ -4,7 +4,8 @@
 ## 0.10.0 — not yet released
 
 * Change key values pairs for additional info's to a dictionary and update the validator to be able to validate the dictionary.
-* Update to latest spec. of delta protocol, specifically: additional properties for all Move/Replace child and annotation commands.
+* Update to latest spec. of delta protocol, specifically: additional properties for all Move/Replace child and annotation commands
+  and simplified properties for MoveAndReplaceInSame[containment|annotation] using an offset. 
 
 
 ## 0.9.0